test(video): split perceptual hash tests into their own file
A module-level pytest.importorskip('numpy') aborts collection of the whole
file, not just the numpy-dependent class, so keeping TestPerceptualHash next
to TestGeometryDistinctness silently skipped the 8 stdlib-only geometry tests
whenever numpy is absent (e.g. the base test CI job).
